requirements.txt
Pillow==6.2.1
prompt-toolkit==1.0.14
Pygments==2.4.2
PyInquirer==1.0.3
regex==2019.11.1
six==1.13.0
wcwidth==0.1.7
